remove unnecessary files

+++ TFT-pytorch/script/prepare_data.py
@@ -1,11 +1,6 @@
 # %% [markdown]
 # # Introduction
-# This file converts the cleaned raw dataset into a single merged file that the TFTModel can work on. The script version available at [prepare_data.py](../script/prepare_data.py).
-# 
-# If you need to change the input feature set, only add that info in the `"data"` section of the json configuration  file. This notebook will update the rest (at least feature column mappings and locations) . If you have pivoted dynamic feature and need to melt that date columns, make sure to keep the feature name as `string` in `"dynamic_features_map"`. If it is already melted and your dynamic file has a `Date` column, `list` or `string` format both is fine.
-# 
-# In the final output all null values are replaced with 0. If you don't want that, comment that out.
-
+# This file converts the cleaned raw dataset into a single merged file that the TFT model can work on. 
 # %% [markdown]
 # # Import libraries
 
@@ -14,58 +9,52 @@ import sys
 sys.path.append( '..' )
 
 # %% [markdown]
-# # Setup storage
-# 
-# You would need the `CovidOct-2022` and `Support files` folders for the dateset. And the `TFT-pytorch` folder for the codes. Upload both of them in the place where you are running the code from. My folder structure looks like this
-# * dataset_raw
-#     * CovidMay17-2022
-#     * Support files
-# * TFT-pytorch
-
-# %% [markdown]
-# ## Googe drive
-# Not needed, since you can run this on CPU. But set `running_on_colab = True` if using. Also update the `cd` path so that it points to the notebook folder in your drive.
-
-# %%
-running_on_colab = False
-
-# if running_on_colab:
-#     from google.colab import drive
-#     drive.mount('/content/drive')
-
-#     %cd /content/drive/My Drive/Projects/Covid/TFT-pytorch/notebooks
-
-# %% [markdown]
 # ## Input
 # If running on colab, modify the below paths accordingly. Note that this config.json is different from the config.json in TF2 folder as that is for the old dataset.
 
 # %%
-from dataclasses import dataclass
 from Class.DataMerger import *
+from argparse import ArgumentParser
 
-@dataclass
-class args:
-    # folder where the cleaned feature file are at
-    dataPath = '../../dataset_raw/CovidMay17-2022'
-    supportPath = '../../dataset_raw/Support files'
-    configPath = '../configurations/age_groups.json'
-    cachePath = None # '../2022_May_cleaned/Total.csv'
+parser = ArgumentParser(description='Prepare Dataset')
 
-    # choose this carefully
-    outputPath = '../2022_May_age_groups/'
+parser.add_argument(
+   '--config', default='../configurations/baseline.json',
+   help='configuration file path'
+)
+
+parser.add_argument(
+   '--input', help='input folder of raw feature files',
+   default='../../dataset_raw/CovidMay17-2022'
+)
+
+parser.add_argument(
+   '--output', default='../2022_May_cleaned',
+   help='output folder for the merged feature file'
+)
+parser.add_argument(
+   '--replace', help='whether to replace the existing features files',
+   action='store_true'
+)
+parser.add_argument(
+   '--support', help='folder of support files (e.g. Population.csv)',
+   default='../../dataset_raw/Support files'
+)
+
+args = parser.parse_args()
 
 # %%
 # create output path if it doesn't exist
-if not os.path.exists(args.outputPath):
-    print(f'Creating output directory {args.outputPath}')
-    os.makedirs(args.outputPath, exist_ok=True)
+if not os.path.exists(args.output):
+    print(f'Creating output directory {args.output}')
+    os.makedirs(args.output, exist_ok=True)
 
 import json
 
 # load config file
-with open(args.configPath) as inputFile:
+with open(args.config) as inputFile:
     config = json.load(inputFile)
-    print(f'Config file loaded from {args.configPath}')
+    print(f'Config file loaded from {args.config}')
     inputFile.close()
 
 # %% [markdown]
@@ -76,17 +65,19 @@ with open(args.configPath) as inputFile:
 
 # %%
 # get merger class
-dataMerger = DataMerger(config, args.dataPath, args.supportPath)
+dataMerger = DataMerger(config, args.input, args.support)
 
 # %%
 # if you have already created the total df one, and now just want to 
 # reuse it to create different population cut
-if args.cachePath:
-    total_df = pd.read_csv(args.cachePath)
+output_path_total = os.path.join(args.output, 'Total.csv') 
+
+# whether to use the cached file
+if (not args.replace) and os.path.exists(output_path_total):
+    total_df = pd.read_csv(output_path_total)
+    print(f'Total.csv already exists in path {output_path_total}. Skipping...')
 else:
     total_df = dataMerger.get_all_features()
-    
-    output_path_total = os.path.join(args.outputPath, 'Total.csv') 
     print(f'Writing total data to {output_path_total}\n')
     
     # rounding up to reduce the file size
@@ -105,7 +96,11 @@ if dataMerger.need_population_cut():
         top_counties = dataMerger.data_config.population_cut[index]
         filename = f"Top_{top_counties}.csv"
 
-        output_path_population_cut = os.path.join(args.outputPath, filename)
+        output_path_population_cut = os.path.join(args.output, filename)
+
+        if (not args.replace) and os.path.exists(output_path_population_cut):
+            print(f'{filename} already exists at {output_path_population_cut}. Skipped.')
+            continue
 
         print(f'Writing top {top_counties} populated counties data to {output_path_population_cut}.')
         population_cuts[index].round(4).to_csv(output_path_population_cut, index=False)
